requirements.txt
youtube-transcript-api>=0.6.0
pytube>=15.0.0
requests>=2.31.0
google-generativeai>=0.3.0
openai-whisper>=20231117
yt-dlp>=2023.11.16